very buggy, but some progress
The buffer isn't working super well for sd reading. I think it has something to do with how the currentFilePosition is being updated. It should be updated so that the file starts seeking from the current_point_idx, wherever that is in the file. It seems like file position should probably be stored in the buffer as well for each point